% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{biooracle_path}
\alias{biooracle_path}
\alias{set_biooracle_root}
\alias{get_biooracle_root}
\alias{make_path}
\title{Get, set and build biooracle paths}
\usage{
set_biooracle_root(root = ".", filename = "~/.biooracle")

get_biooracle_root(filename = "~/.biooracle")

biooracle_path(..., root = get_biooracle_root())

make_path(path)
}
\arguments{
\item{root}{chr the root biooracle data path}

\item{filename}{chr the filename where we store the path}

\item{...}{file path segments passed to \link{file.path}}
}
\value{
a file or directory path description
}
\description{
Get, set and build biooracle paths
}
