% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{biooracle_url}
\alias{biooracle_url}
\title{Retrieve the current Bio-Oracle URL for griddap}
\usage{
biooracle_url()
}
\value{
character URL
}
\description{
Retrieve the current Bio-Oracle URL for griddap
}
