% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{compose_database}
\alias{compose_database}
\title{Given a stars object and an dataset_id, compose a small database
of the files that might be generated.}
\usage{
compose_database(x, id)
}
\arguments{
\item{x}{stars object}

\item{id}{chr, a dataset_id like "thetao_ssp119_2020_2100_depthmin"}
}
\value{
a tabular database with scenario, year, z, param and trt
}
\description{
Given a stars object and an dataset_id, compose a small database
of the files that might be generated.
}
