% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{compose_filename}
\alias{compose_filename}
\title{Compose a filename from a database}
\usage{
compose_filename(x, path = NULL, ext = ".tif")
}
\arguments{
\item{x}{a table database}

\item{path}{chr, the directory}

\item{ext}{chr, the extension}
}
\value{
character filename(s)
}
\description{
Compose a filename from a database
}
