% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{decompose_filename}
\alias{decompose_filename}
\title{Decompose a filename into a table database}
\usage{
decompose_filename(x, ext = ".tif")
}
\arguments{
\item{x}{chr one or more filenames}

\item{ext}{chr, the extension to strip}
}
\value{
tabular database
}
\description{
Decompose a filename into a table database
}
