% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_biooracle}
\alias{download_biooracle}
\title{Downloads a file using \link[biooracler:download_layers]{biooracler::download_layers}}
\usage{
download_biooracle(
  dataset_id = "thetao_ssp119_2020_2100_depthmin",
  constraints = NULL,
  base_url = "https://erddap.bio-oracle.org/erddap/griddap",
  temp_path = biooracle_path("temp")
)
}
\arguments{
\item{dataset_id}{chr, see \link[biooracler:download_layers]{biooracler::download_layers}}

\item{constraints}{chr, a formatted constraints string}

\item{base_url}{chr the serbvert URL for griddap}

\item{temp_path}{chr, the temporary data path}
}
\value{
the fully qualified downloaded file name
}
\description{
Downloads a file using \link[biooracler:download_layers]{biooracler::download_layers}
}
