% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{fetch_biooracle}
\alias{fetch_biooracle}
\title{Fetch a Bio-Oracle file.}
\usage{
fetch_biooracle(
  dataset_id = "thetao_ssp119_2020_2100_depthmin",
  vars = NULL,
  time = NULL,
  bb = NULL,
  ...
)
}
\arguments{
\item{dataset_id}{chr, see \link[biooracler:download_layers]{biooracler::download_layers}}

\item{vars}{NULL or chr, list of variable names (defaults to all)}

\item{time}{NULL, POSIXct or Date min and max, (defaults to all)}

\item{bb}{bounding box (or object from which \link[sf:st_bbox]{sf::st_bbox} can be derived)
(defaults to max spatial range)}

\item{...}{other arguments for download_biooracle}
}
\value{
the name of the file retrieved
}
\description{
This is a convenience wrapper around \link{download_biooralce}.
}
