% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{fetch_terrain}
\alias{fetch_terrain}
\title{Fetch terrain data and possibly archive}
\usage{
fetch_terrain(
  archive = TRUE,
  data_dir = biooracle_path("nwa"),
  bb = c(xmin = -77, xmax = -51.5, ymin = 37.9, ymax = 56.7)
)
}
\arguments{
\item{archive}{logical, if TRUE save into a \code{terrain} subdirectory of \code{path}}

\item{data_dir}{chr, the output path}

\item{bb}{a 4 element bounding box}
}
\value{
stars object
}
\description{
Fetch terrain data and possibly archive
}
