% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{generate_constraints}
\alias{generate_constraints}
\title{Generate constrints query gicen the dataset_id, variables, time and bounding box}
\usage{
generate_constraints(
  dataset_id = "thetao_ssp119_2020_2100_depthmin",
  vars = NULL,
  time = NULL,
  bb = NULL
)
}
\arguments{
\item{dataset_id}{char, the dataset identifier}

\item{vars}{NULL or chr, list of variable names (defaults to all)}

\item{time}{NULL, POSIXct or Date min and max, (defaults to all)}

\item{bb}{bounding box (or object from which \link[sf:st_bbox]{sf::st_bbox} can be derived)
(defaults to max spatial range)}
}
\value{
the complete query string
}
\description{
Generate constrints query gicen the dataset_id, variables, time and bounding box
}
