% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biooracle.R
\name{list_layers}
\alias{list_layers}
\title{List the layers available - requires biooracle}
\usage{
list_layers()
}
\value{
tibble with the available layer information
}
\description{
List the layers available - requires biooracle
}
