% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{read_terrain}
\alias{read_terrain}
\title{Read one of more terrain rasters}
\usage{
read_terrain(
  what = c("area", "aspect", "bathymetry_max", "bathymetry_mean", "bathymetry_min",
    "coastline", "landmass", "slope", "terrain_ruggedness_index",
    "topographic_position_index"),
  path = biooracle_path("nwa"),
  point = NA,
  crs = 4326
)
}
\arguments{
\item{what}{chr one or more terrain layers}

\item{path}{the path to the regional dataset}

\item{point}{NA or logical to alter the \code{point} dimension attribute}

\item{crs}{st_crs or numeric or char, the CRS desired}
}
\value{
stars object
}
\description{
Read one of more terrain rasters
}
