% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{set_point_attr}
\alias{set_point_attr}
\title{Set the value of a stars object's \code{point} dimension attribute}
\usage{
set_point_attr(x, point = c(NA, TRUE, FALSE)[1])
}
\arguments{
\item{x}{stars object}

\item{point}{NA or logical}
}
\value{
the updated stars object
}
\description{
Set the value of a stars object's \code{point} dimension attribute
}
