% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_biooracle}
\alias{write_biooracle}
\title{Write a Bio-Oracle stars object guided by a database table}
\usage{
write_biooracle(x, db, path = ".")
}
\arguments{
\item{x}{stars object with one or more params at one or more times}

\item{db}{biooralce database}

\item{path}{chr, the data path}
}
\value{
the input \code{x} object (suitable for piping)
}
\description{
Write a Bio-Oracle stars object guided by a database table
}
