% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{write_database}
\alias{write_database}
\title{Write the file-list database}
\usage{
write_database(x, path, filename = "database", append = FALSE)
}
\arguments{
\item{x}{the tibble or data.frame database}

\item{path}{character the directory to where the database should reside}

\item{filename}{character, the name of the database file}

\item{append}{logical, if TRUE try to append to the existing database}
}
\value{
a tibble
}
\description{
Write the file-list database
}
