% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{archive_biooracle}
\alias{archive_biooracle}
\title{Archive a downloaded Bio-Oracle netcdf file (see \link{fetch_biooracle}).}
\usage{
archive_biooracle(x, path = ".", dataset_id = NULL)
}
\arguments{
\item{x}{stars or chr a stars object or the name of the file to archive}

\item{path}{chr the output path}

\item{dataset_id}{chr, only required if \code{x} is a stars object}
}
\value{
a small database (table) of metadata
}
\description{
Archiving involves writing each layer to (date, depth and variable slice) to a
regional subfolder.  Files are organized into decadal subdirectories.  Metadata
to help navigate the archive are organized into the \code{database} file.  See
\link{read_database} for details.
}
